requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
Werkzeug==2.3.7